Puzzle Implementation and Parkour
Gave the torch puzzle wall textures.
Added a sliding floor tile that reveals the entrance to the boss room when two key items are acquired by the player.
Key items can be acquired by completing puzzles. 1 is at the top of the library parkour, and 1 is spawned after the symmetry puzzle is completed.
Changed the default number of jumps to 2, so as to make the library parkour possible.